--- src/pottery-web-app/src/404Page.tsx
@@ -1,7 +1,0 @@
-import { CSSProperties } from "react";
-
-const NotFound = () => {
-  return <div style={{ color: "red" }}>404 Page not Found</div>;
-};
-
-export default NotFound;
