+++ src/pottery-web-app/src/components/banner.tsx
@@ -23,7 +23,9 @@ const Banner = () => {
           <button
             onClick={() => {
               localStorage.removeItem("token");
+              localStorage.removeItem("userId");
               setUsername(null);
+              navigate("/");
             }}
           >
             Logout
