+++ src/pottery-web-app/src/pages/login/Login.tsx
@@ -27,7 +27,9 @@ const Login = () => {
   };
 
   useEffect(() => {
-    isUserAuth().then((res) => (res.data.isLoggedIn ? navigate("/") : null));
+    isUserAuth().then((res) =>
+      res.data.isLoggedIn ? navigate("/pottery") : null
+    );
   }, [msg]);
 
   return (
