+++ src/pottery-web-app/src/pages/potNotes/FormValidationSchema.tsx
