+++ src/pottery-web-app/src/pages/potNotes/parts/formSections/GeneralSection.tsx
@@ -1,5 +1,5 @@
 import { Form, Row, Col } from "react-bootstrap";
-import { CATEGORY_OPTIONS, STAGE_OPTIONS } from "../../common/Constants";
+import { CATEGORY_OPTIONS, STAGE_OPTIONS } from "../../utility/Constants";
 import { useFormContext, Controller } from "react-hook-form";
 import { ErrorMessage } from "@hookform/error-message";
