+++ src/pottery-web-app/src/pages/potNotes/utility/utilityFunctions.tsx
