+++ src/pottery-web-app/src/pages/pottery/GalleryHeader.tsx
