+++ src/pottery-web-app/src/pages/pottery/main.tsx
@@ -13,7 +13,6 @@ const Main = () => {
   }, []);
 
   const fetchPots = (): void => {
-    console.log(localStorage.getItem("userId"));
     getPots(localStorage.getItem("userId")).then(
       ({ data: { pots } }: IPotInfo[] | any) => {
         setPots(pots);
